Script 5 changes
Changes have been made to the Clustering script, as the naming conventions in the base code were outdated. All control files have also been updated to match naming conventions in the Clustering script and the Bolivian case. The Script is run twice, one for solarpv and one for wind (configured in the control files).